// src/pages/ProgramReconstruction/types/types.ts
export interface IAddStageValues {
	name: string;
	description: string;
}

export interface IEditStageValues {
	name: string;
	description: string;
}

export interface IAddProcessValues {
	name: string;
	description: string;
	result: string;
}

export interface IEditProcessValues {
	name: string;
	description: string;
	result: string;
}